+++ Modules/Inventario/Http/Controllers/UbicacionesHistorialController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Modules\Inventario\Http\Controllers;
+
+use Illuminate\Routing\Controller;
+
+class UbicacionesHistorialController extends Controller
+{
+    public function index()
+    {
+        return view('inventario::ubicaciones.historial');
+    }
+}